requirements.txt
playwright~=1.53.0
pytest~=8.4.1
python-dotenv~=1.1.1